refactor: extract internal/senml package from internal/sensors
Moves SenML parsing (pure encoding, no domain deps) to its own package.
ReadingsService now imports senml.Parse instead of the local ParseSenML.

Closes #96
